refactor: extracted bar chart code to utils
